# docs/api-reference/openapi.json
{
  "info": {
    "description": "Generated from the live Flask url_map \u2014 see potato/server_utils/openapi_spec.py.\n\nMost endpoints are session-authenticated: POST /auth with `username` and `password` as form data, then reuse the session cookie. `x-potato-auth` records the decorators guarding an operation, and `x-potato-requires-config` marks operations that only exist when the named config is enabled.\n\nResponse schemas are intentionally not modelled; this spec is an authoritative index of what exists, not a payload contract.",
    "title": "Potato HTTP API",
    "version": "2.8.2"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/admin": {
      "get": {
        "operationId": "get_admin",
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Only available when the server runs with debug: true"
          }
        },
        "summary": "Serve the admin dashboard page.",
        "tags": [
          "core"
        ],
        "x-potato-auth": [
          "inline:admin_api_key",
          "inline:debug_only"
        ]
      }
    },
...
    "/admin/api/embedding_viz/data": {
      "get": {
        "operationId": "get_admin_api_embedding_viz_data",
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Admin role required"
          }
        },
        "summary": "Get visualization data for the embedding scatter plot.",
        "tags": [
          "core"
        ],
        "x-potato-auth": [
          "inline:admin_access"
        ]
      }
    },
...
    "/api/rollout/set": {
      "get": {
        "operationId": "get_rollout_set",
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Authentication required"
          }
        },
        "summary": "Build and return the current item's rollout manifest.",
        "tags": [
          "core"
        ],
        "x-potato-auth": [
          "inline:session_username"
        ]
      }
    },
...
    "/datasets": {
      "get": {
        "operationId": "get_datasets_index",
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Admin role required"
          }
        },
        "summary": "The bare prefix, which is the URL the boot log's \"Registered",
        "tags": [
          "datasets"
        ],
        "x-potato-auth": [
          "admin_required"
        ],
        "x-potato-requires-config": "datasets.enabled: true"
      }
    },
    "/datasets/": {
      "get": {
        "operationId": "get_datasets_index__datasets",
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Admin role required"
          }
        },
        "summary": "The bare prefix, which is the URL the boot log's \"Registered",
        "tags": [
          "datasets"
        ],
        "x-potato-auth": [
          "admin_required"
        ],
        "x-potato-requires-config": "datasets.enabled: true"
      }
    },
    "/datasets/admin": {
      "get": {
        "operationId": "get_datasets_admin",
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Admin role required"
          }
        },
        "tags": [
          "datasets"
        ],
        "x-potato-auth": [
          "admin_required"
        ],
        "x-potato-requires-config": "datasets.enabled: true"
      }
    },
    "/datasets/api/datasets": {
      "get": {
        "operationId": "get_datasets_api_list_datasets",
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Admin role required"
          }
        },
        "tags": [
          "datasets"
        ],
        "x-potato-auth": [
          "admin_required"
        ],
        "x-potato-requires-config": "datasets.enabled: true"
      },
      "post": {
        "operationId": "post_datasets_api_create_dataset",
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Admin role required"
          }
        },
        "tags": [
          "datasets"
        ],
        "x-potato-auth": [
          "admin_required"
        ],
        "x-potato-requires-config": "datasets.enabled: true"
      }
    },
    "/datasets/api/datasets/{name}": {
      "delete": {
        "operationId": "delete_datasets_api_delete_dataset",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Admin role required"
          }
        },
        "tags": [
          "datasets"
        ],
        "x-potato-auth": [
          "admin_required"
        ],
        "x-potato-requires-config": "datasets.enabled: true"
      },
      "get": {
        "operationId": "get_datasets_api_get_dataset",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Admin role required"
          }
        },
        "tags": [
          "datasets"
        ],
        "x-potato-auth": [
          "admin_required"
        ],
        "x-potato-requires-config": "datasets.enabled: true"
      }
    },
    "/datasets/api/datasets/{name}/examples": {
      "get": {
        "operationId": "get_datasets_api_list_examples",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Admin role required"
          }
        },
        "tags": [
          "datasets"
        ],
        "x-potato-auth": [
          "admin_required"
        ],
        "x-potato-requires-config": "datasets.enabled: true"
      },
      "post": {
        "operationId": "post_datasets_api_add_examples",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Admin role required"
          }
        },
        "tags": [
          "datasets"
        ],
        "x-potato-auth": [
          "admin_required"
        ],
        "x-potato-requires-config": "datasets.enabled: true"
      }
    },
...
    "/datasets/api/datasets/{name}/tag": {
      "post": {
        "operationId": "post_datasets_api_tag_version",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Admin role required"
          }
        },
        "tags": [
          "datasets"
        ],
        "x-potato-auth": [
          "admin_required"
        ],
        "x-potato-requires-config": "datasets.enabled: true"
      }
    },
...
    "/datasets/experiments/compare": {
      "get": {
        "operationId": "get_datasets_compare_experiments",
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Admin role required"
          }
        },
        "tags": [
          "datasets"
        ],
        "x-potato-auth": [
          "admin_required"
        ],
        "x-potato-requires-config": "datasets.enabled: true"
      }
    },
    "/datasets/view/{name}": {
      "get": {
        "operationId": "get_datasets_view_dataset",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Admin role required"
          }
        },
        "tags": [
          "datasets"
        ],
        "x-potato-auth": [
          "admin_required"
        ],
        "x-potato-requires-config": "datasets.enabled: true"
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:8000"
    }
  ]
}